Crate street_network

Re-exports

pub use self::geometry::InputRoad;

Modules

edit 🔒

export 🔒

geometry

OSM describes roads as center-lines that intersect. Turn these into road and intersection polygons roughly by

initial

Naming is confusing, but StreetNetwork -> InitialMap -> Map. InitialMap is separate pretty much just for the step of producing https://a-b-street.github.io/docs/tech/map/importing/geometry.html.

lane_specs 🔒

osm

Useful utilities for working with OpenStreetMap.

transform 🔒

types 🔒

Structs

LaneSpec

MapConfig

NamePerLanguage

None corresponds to the native name

OriginalRoad

A way to refer to roads across many maps and over time. Also trivial to relate with OSM to find upstream problems.

RawIntersection

RawRoad

StreetNetwork

TurnRestriction

Enums

BufferType

Direction

DrivingSide

IntersectionType

LaneType

RestrictionType

Constants

NORMAL_LANE_THICKNESS

SIDEWALK_THICKNESS

Functions

get_lane_specs_ltr

intersection_polygon